Autocomplete: Move context logic into a seperate folder and improve naming consistency
