+++ vscode/src/completions/context/bestJaccardMatch.test.ts
