+++ vscode/src/completions/context/context.ts
@@ -2,17 +2,11 @@ import * as vscode from 'vscode'
 
 import { CodebaseContext } from '@sourcegraph/cody-shared/src/codebase-context'
 
+import { ContextSnippet } from '../types'
+
 import { getContextFromEmbeddings } from './context-embeddings'
 import { getContextFromCurrentEditor } from './context-local'
 import { DocumentHistory } from './history'
-
-/**
- * Keep property names in sync with the `EmbeddingsSearchResult` type.
- */
-export interface ReferenceSnippet {
-    fileName: string
-    content: string
-}
 
 export interface GetContextOptions {
     document: vscode.TextDocument
@@ -26,7 +20,7 @@ export interface GetContextOptions {
 }
 
 export interface GetContextResult {
-    context: ReferenceSnippet[]
+    context: ContextSnippet[]
     logSummary: {
         embeddings?: number
         local?: number
@@ -50,9 +44,9 @@ export async function getContext(options: GetContextOptions): Promise<GetContext
      * Discard editor matches for files with embedding matches.
      */
     const usedFilenames = new Set<string>()
-    const context: ReferenceSnippet[] = []
+    const context: ContextSnippet[] = []
     let totalChars = 0
-    function addMatch(match: ReferenceSnippet): boolean {
+    function addMatch(match: ContextSnippet): boolean {
         // TODO(@philipp-spiess): We should de-dupe on the snippet range and not
         // the file name to allow for more than one snippet of the same file
         if (usedFilenames.has(match.fileName)) {
