+++ vscode/src/completions/context/history.ts
