+++ vscode/src/completions/providers/unstable-codegen.ts
@@ -1,8 +1,7 @@
 import fetch from 'isomorphic-fetch'
 
 import { logger } from '../../log'
-import { ReferenceSnippet } from '../context'
-import { Completion } from '../types'
+import { Completion, ContextSnippet } from '../types'
 import { isAbortError } from '../utils'
 
 import { Provider, ProviderConfig, ProviderOptions } from './provider'
@@ -21,7 +20,7 @@ export class UnstableCodeGenProvider extends Provider {
         this.serverEndpoint = unstableCodeGenOptions.serverEndpoint
     }
 
-    public async generateCompletions(abortSignal: AbortSignal, snippets: ReferenceSnippet[]): Promise<Completion[]> {
+    public async generateCompletions(abortSignal: AbortSignal, snippets: ContextSnippet[]): Promise<Completion[]> {
         const params = {
             debug_ext_path: 'cody',
             lang_prefix: `<|${mapVSCodeLanguageIdToModelId(this.options.languageId)}|>`,
@@ -121,7 +120,7 @@ interface Context {
     }[]
 }
 
-function prepareContext(snippets: ReferenceSnippet[], fileName: string): Context {
+function prepareContext(snippets: ContextSnippet[], fileName: string): Context {
     const windows: Context['windows'] = []
 
     // the model expects a similarly to rank the order and priority to insert
