+++ vscode/src/completions/tracer/index.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode'
 
 import { CompletionParameters } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/types'
 
-import { GetContextResult } from '../context'
+import { GetContextResult } from '../context/context'
 import { InlineCompletionsResult } from '../getInlineCompletions'
 import { CompletionProviderTracerResultData, Provider } from '../providers/provider'
