+++ vscode/src/completions/types.ts
@@ -12,3 +12,11 @@ export interface InlineCompletionItem {
     insertText: string
     range?: Range
 }
+
+/**
+ * Keep property names in sync with the `EmbeddingsSearchResult` type.
+ */
+export interface ContextSnippet {
+    fileName: string
+    content: string
+}
