+++ vscode/test/completions/run-code-completions-on-dataset.ts
@@ -11,8 +11,8 @@ import { NoopEditor } from '@sourcegraph/cody-shared/src/editor'
 import { SourcegraphNodeCompletionsClient } from '@sourcegraph/cody-shared/src/sourcegraph-api/completions/nodeClient'
 import { NOOP_TELEMETRY_SERVICE } from '@sourcegraph/cody-shared/src/telemetry'
 
-import { GetContextResult } from '../../src/completions/context'
-import { VSCodeDocumentHistory } from '../../src/completions/history'
+import { GetContextResult } from '../../src/completions/context/context'
+import { VSCodeDocumentHistory } from '../../src/completions/context/history'
 import { createProviderConfig } from '../../src/completions/providers/createProvider'
 import { InlineCompletionItemProvider } from '../../src/completions/vscodeInlineCompletionItemProvider'
 import { getFullConfig } from '../../src/configuration'
